% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{geno}
\alias{geno}
\alias{geno<-}
\alias{geno,QuasarExperiment-method}
\alias{geno<-,QuasarExperiment-method}
\title{Access the lazy genotype matrix}
\usage{
geno(x)

geno(x) <- value

\S4method{geno}{QuasarExperiment}(x)

\S4method{geno}{QuasarExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{QuasarExperiment}.}

\item{value}{A \link[BEDMatrix:BEDMatrix]{BEDMatrix::BEDMatrix} or matrix-like object
(samples × variants).}
}
\value{
\code{geno()} returns the \code{geno} slot (a \code{BEDMatrix}, samples × variants).
}
\description{
Access the lazy genotype matrix
}
\examples{
exdir <- system.file("extdata", package = "QuasarExperiment")
qe <- QuasarExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed"),
    covFile     = file.path(exdir, "cov-n100.tsv")
)
geno(qe)
geno(qe)[1:3, 1:5]
}
