% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{plinkPrefix}
\alias{plinkPrefix}
\alias{plinkPrefix<-}
\alias{plinkPrefix,QuasarExperiment-method}
\alias{plinkPrefix<-,QuasarExperiment-method}
\title{Access the PLINK file prefix}
\usage{
plinkPrefix(x)

plinkPrefix(x) <- value

\S4method{plinkPrefix}{QuasarExperiment}(x)

\S4method{plinkPrefix}{QuasarExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{QuasarExperiment}.}

\item{value}{A length-one character string.}
}
\value{
\code{plinkPrefix()} returns the path prefix for the \code{.bed}/\code{.bim}/\code{.fam}
file set.
}
\description{
Access the PLINK file prefix
}
