% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,QuasarExperiment-method}
\alias{show,QuasarExperiment-method}
\title{Show a QuasarExperiment}
\usage{
\S4method{show}{QuasarExperiment}(object)
}
\arguments{
\item{object}{A \link{QuasarExperiment}.}
}
\description{
Displays a compact summary of the object dimensions and key slots.
}
