% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{variantRanges}
\alias{variantRanges}
\alias{variantRanges<-}
\alias{variantRanges,QuasarExperiment-method}
\alias{variantRanges<-,QuasarExperiment-method}
\title{Access variant genomic ranges}
\usage{
variantRanges(x)

variantRanges(x) <- value

\S4method{variantRanges}{QuasarExperiment}(x)

\S4method{variantRanges}{QuasarExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{QuasarExperiment}.}

\item{value}{A \link[GenomicRanges:GRanges]{GenomicRanges::GRanges} with length equal to \code{nrow(geno(x))}.}
}
\value{
\code{variantRanges()} returns a \code{GRanges} with one range per variant.
Metadata columns include \code{snp_id}, \code{ref}, and \code{alt}.
}
\description{
Access variant genomic ranges
}
\examples{
exdir <- system.file("extdata", package = "QuasarExperiment")
qe <- QuasarExperiment(
    plinkPrefix = file.path(exdir, "chr22-n100"),
    phenoFile   = file.path(exdir, "mean-pheno-n100.bed"),
    covFile     = file.path(exdir, "cov-n100.tsv")
)
variantRanges(qe)
}
