% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{geno}
\alias{geno}
\alias{geno<-}
\alias{geno,QuasarExperiment-method}
\alias{geno<-,QuasarExperiment-method}
\title{Access the lazy genotype matrix}
\usage{
geno(x)

geno(x) <- value

\S4method{geno}{QuasarExperiment}(x)

\S4method{geno}{QuasarExperiment}(x) <- value
}
\arguments{
\item{x}{A \link{QuasarExperiment}.}

\item{value}{A \link[BEDMatrix:BEDMatrix]{BEDMatrix::BEDMatrix} or matrix-like object
(variants × samples).}
}
\value{
\code{geno()} returns the \code{geno} slot (a \code{BEDMatrix} or matrix-like).
}
\description{
Access the lazy genotype matrix
}
\examples{
# geno(qe)          # BEDMatrix: 69638 x 100
# geno(qe)[1:5, ]   # materialize five variant rows
}
