% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{regionHits}
\alias{regionHits}
\title{Extract region-level results from a quasar run}
\usage{
regionHits(x)
}
\arguments{
\item{x}{The list returned by \code{\link[=runQuasar]{runQuasar()}}.}
}
\value{
A \link[GenomicRanges:GRanges]{GenomicRanges::GRanges} with one entry per phenotype feature,
containing metadata columns \code{feature_id} and \code{pvalue}.
}
\description{
Extract region-level results from a quasar run
}
\seealso{
\code{\link[=runQuasar]{runQuasar()}}, \code{\link[=variantHits]{variantHits()}}
}
